Fix broken BuffishShorcut command so it accepts an argument
It wasn't properly accepting the optional argument of the shortcut you'd
like to follow. Updated the readme to elaborate on the feature.

Also changed the initialization of the plugin to the `.setup()` pattern
that is most common right now, though it is currently auto setting up
and not accepting any configuration.